updated test/unit structure
